server/requirements.txt
Flask
gunicorn
mysql-connector-python
python-dotenv
openpyxl
